refactor: crypto utils typescript
